🚚 rename: 스택 구조 변경
