+++ navigation/ListStack.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+import ListScreen from "../screens/ListScreen";
+import SearchScreen from "../screens/SearchScreen";
+
+const Stack = createStackNavigator();
+
+export default function ListStack() {
+  return (
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="List" component={ListScreen} />
+      <Stack.Screen name="Search" component={SearchScreen} />
+    </Stack.Navigator>
+  );
+}
