+++ navigation/SettingsStack.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+import SettingsScreen from "../screens/SettingsScreen";
+
+const Stack = createStackNavigator();
+
+export default function SettingsStack() {
+  return (
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="Settings" component={SettingsScreen} />
+      {/* <Stack.Screen name="AccountSettings" component={AccountSettingsScreen} /> */}
+    </Stack.Navigator>
+  );
+}
