+++ screens/SearchScreen.js
@@ -1,0 +1,13 @@
+import { Text, View } from "react-native";
+import React from "react";
+import Layout from "../layouts/Layout";
+
+const SearchScreen = () => {
+  return (
+    <Layout>
+      <Text>SearchScreen</Text>
+    </Layout>
+  );
+};
+
+export default SearchScreen;
