+++ screens/SettingsScreen.js
@@ -1,0 +1,13 @@
+import { Text, View } from "react-native";
+import React from "react";
+import Layout from "../layouts/Layout";
+
+const SettingsScreen = () => {
+  return (
+    <Layout>
+      <Text>SettingsScreen</Text>
+    </Layout>
+  );
+};
+
+export default SettingsScreen;
